move common unit tests to common_unit_tests.py
